........
  r33154 | sueastside | 2009-11-10 23:42:22 +0000 (Tue, 10 Nov 2009) | 1 line
  
  Added RegressionTest, an application to compare stock rendered meshes/worlds with local rendering.
........
  r33155 | crystal-manage | 2009-11-11 04:01:44 +0000 (Wed, 11 Nov 2009) | 8 lines
  
  master header files: Automated master header file repair.
  Visual-C++ project files: Automated Visual-C++ project file repair.
  Texinfo files: Automated Texinfo @node and @menu repair.
  Swig Python files: Automated Swig Python file repair.
  Swig Perl5 files: Automated Swig Perl5 file repair.
  User's Manual: Automated Texinfo to HTML conversion.
  csver.h: Automated csver.h regeneration.
........
  r33156 | sueastside | 2009-11-11 20:51:33 +0000 (Wed, 11 Nov 2009) | 1 line
  
  Renamed regressiontest to smoketest as it isn't in-depth enough.
........
  r33158 | crystal-manage | 2009-11-12 04:11:25 +0000 (Thu, 12 Nov 2009) | 8 lines
  
  master header files: Automated master header file repair.
  Visual-C++ project files: Automated Visual-C++ project file repair.
  Texinfo files: Automated Texinfo @node and @menu repair.
  Swig Python files: Automated Swig Python file repair.
  Swig Perl5 files: Automated Swig Perl5 file repair.
  User's Manual: Automated Texinfo to HTML conversion.
  csver.h: Automated csver.h regeneration.
........
  r33164 | res2002 | 2009-11-13 16:09:54 +0000 (Fri, 13 Nov 2009) | 1 line
  
  gcc warning fix
........
  r33165 | res2002 | 2009-11-13 16:15:44 +0000 (Fri, 13 Nov 2009) | 1 line
  
  Make dtor virtual
........
  r33166 | res2002 | 2009-11-13 18:59:56 +0000 (Fri, 13 Nov 2009) | 2 lines
  
  Detect how to turn linker warnings into errors.
  Use that to disable use of the -E linker option which emits a warning on binutils 2.20 on MinGW.
........
  r33167 | res2002 | 2009-11-13 19:03:45 +0000 (Fri, 13 Nov 2009) | 1 line
  
  Regenerated configure
........
  r33170 | res2002 | 2009-11-13 21:10:33 +0000 (Fri, 13 Nov 2009) | 3 lines
  
  Change Swig rule to optionally take the name of a header file as an argument.
  This allows the Swig rule to properly freeze a generated header file.
........
  r33175 | res2002 | 2009-11-14 01:06:56 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Move shader snippets to their own dir to avoid them being mistaken for actual shaders
........
  r33176 | res2002 | 2009-11-14 01:21:35 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Fix botched XML
........
  r33177 | res2002 | 2009-11-14 02:08:41 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Remove <attribute> (doesn't really work); Fix input name mismatch
........
  r33178 | crystal-manage | 2009-11-14 04:00:16 +0000 (Sat, 14 Nov 2009) | 8 lines
  
  master header files: Automated master header file repair.
  Visual-C++ project files: Automated Visual-C++ project file repair.
  Texinfo files: Automated Texinfo @node and @menu repair.
  Swig Python files: Automated Swig Python file repair.
  Swig Perl5 files: Automated Swig Perl5 file repair.
  User's Manual: Automated Texinfo to HTML conversion.
  csver.h: Automated csver.h regeneration.
........
  r33179 | Khakilord | 2009-11-14 16:21:50 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Fixes to actor collision when jumping and for maps with geometry bugs on the floor. Also changed surface normal requirement to be y: 0.7 so steep slopes again can't be climbed.
........
  r33180 | res2002 | 2009-11-14 16:23:56 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Fix formula for multiplicative detail (by Christof 'kinea' Hagedorn)
........
  r33181 | res2002 | 2009-11-14 16:36:02 +0000 (Sat, 14 Nov 2009) | 1 line
  
  Fix snippets path for 'old' shaders
........


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/occluvis@33288 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/threading/win32_thread.cpp
@@ -137,7 +137,7 @@ namespace Implementation
       startupBarrier->Wait ();
       
       // Set the name, for debugging
-      SetThreadName (-1, runnable->GetName ());
+      SetThreadName ((DWORD)-1, runnable->GetName ());
 
       // Run      
       runnable->Run ();
